24-10-2015_01
Upgrade System implemented
- although, ammo upgrades don't work (no ammo system yet!)

- player starts out with 10k squarebucks to play around

- movement speed, attack speed, and damage can be upgraded

- money updates when you use it!!! hory shit!!!

+++ Assets/Scripts/GameScene/CombatModule.cs
@@ -17,7 +17,7 @@ public class CombatModule : MonoBehaviour
 
     }
 
-    public void DealDamage(int a_damage, GameObject a_target)
+    public void DealDamage(float a_damage, GameObject a_target)
     {
         if (a_target.GetComponent<CombatModuleAttributes>())
         {
